pyproject.toml
[tool.poetry]
name = "mass-watermark-stamper"
version = "2.0.2"
description = "A tool for automatically cropping all images in a given folder, resizing, reshaping them to a square and applying a watermark."
authors = ["Nyaaa <1457047+Nyaaa@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
pillow = "^9.2.0"
gooey = "^1.0.8.1"


[tool.poetry.group.dev.dependencies]
pyinstaller = "^5.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
